refactor: group gameplay and runner modules by directory
