+++ src/simulation/runner/phases/index.ts
@@ -1,0 +1,4 @@
+export { runMovementPhase } from './movement';
+export { runAttackPhase } from './weaponAttack';
+export { runPhysicalAttackPhase } from './physicalAttack';
+export { runPSRPhase, runHeatPhase } from './postCombat';
