+++ src/simulation/runner/phases/utils.ts
@@ -1,7 +1,7 @@
 import { GameEventType, GamePhase, IGameEvent } from '@/types/gameplay';
 import { D6Roller } from '@/utils/gameplay/hitLocation';
 
-import { SeededRandom } from '../core/SeededRandom';
+import { SeededRandom } from '../../core/SeededRandom';
 
 export function createD6Roller(random: SeededRandom): D6Roller {
   return () => Math.floor(random.next() * 6) + 1;
