+++ src/utils/gameplay/gameState/actionLocking.ts
