+++ src/utils/gameplay/gameState/extendedCombat.ts
