+++ src/utils/gameplay/gameState/initialization.ts
