+++ src/utils/gameplay/physicalAttacks.ts
@@ -1,7 +1,1 @@
-export * from './physicalAttacksConstants';
-export * from './physicalAttacksDamage';
-export * from './physicalAttacksDecision';
-export * from './physicalAttacksResolution';
-export * from './physicalAttacksRestrictions';
-export * from './physicalAttacksToHit';
-export * from './physicalAttacksTypes';
+export * from './physicalAttacks/index';
