+++ src/utils/gameplay/physicalAttacks/constants.ts
